rename to underscore generality
