+++ src/main.rs
@@ -1,9 +1,9 @@
-mod clinical_status_manager;
 mod contact;
 mod infection_propagation_loop;
 mod infectiousness_manager;
 mod parameters;
 mod population_loader;
+mod property_progression_manager;
 pub mod rate_fns;
 mod settings;
 mod symptom_progression;
