+++ src/property_progression_manager.rs
